refactoring otp generator class
