+++ src/main/java/com/wallet/shieldpay/controllers/ShieldPayUserController.java
@@ -1,0 +1,10 @@
+package com.wallet.shieldpay.controllers;
+
+import jakarta.validation.Valid;
+
+public class ShieldPayUserController {
+
+    public void register(@Valid String string){
+
+    }
+}
